pyproject.toml
[project]
name = "huggingface-agents-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "smolagents[litellm]>=1.24.0",
    "transformers>=4.57.6",
]
